Refactor: 기업, 유저 dto 분리
