+++ bin/main/mapper/user.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="shop.mtcoding.job.model.user.UserRepository">
     <insert id="insert">
-      insert into user_tb(username, password, name, email, contact, created_at) 
-      values(#{username}, #{password}, #{name}, #{email}, #{contact}, now()) 
+      insert into user_tb(username, password, salt, name, email, contact, profile, created_at) 
+      values(#{username}, #{password}, #{salt}, #{name}, #{email}, #{contact}, #{profile}, now()) 
    </insert>
     
    <update id="updateById">
