+++ src/main/java/shop/mtcoding/job/dto/user/UserReqDto.java
@@ -13,33 +13,14 @@ public class UserReqDto {
 
     @Setter
     @Getter
-    public static class LoginEnterpriseReqDto {
-        private String enterpriseName;
-        private String password;
-    }
-
-    @Setter
-    @Getter
     public static class JoinUserReqDto {
         private String username;
         private String password;
+        private String salt;
         private String name;
         private String email;
         private String contact;
         private String profile;
-    }
-
-    @Setter
-    @Getter
-    public static class JoinEnterpriseReqDto {
-        private String enterpriseName;
-        private String password;
-        private String address;
-        private String contact;
-        private String image;
-        private String email;
-        private String sector;
-        private String size;
     }
 
 }
