+++ src/main/java/shop/mtcoding/job/model/enterprise/EnterpriseRepository.java
@@ -3,6 +3,7 @@ package shop.mtcoding.job.model.enterprise;
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface EnterpriseRepository {
@@ -10,7 +11,13 @@ public interface EnterpriseRepository {
 
         public Enterprise findById(int id);
 
-        public int insert(Enterprise enterprise);
+        public int insert(@Param("enterpriseName") String enterpriseName, @Param("password") String password,
+                        @Param("salt") String salt,
+                        @Param("address") String address,
+                        @Param("contact") String contact,
+                        @Param("image") String image,
+                        @Param("email") String email, @Param("size") String size,
+                        @Param("sector") String sector);
 
         public int updateById(Enterprise enterprise);
