+++ src/main/java/shop/mtcoding/job/service/UserService.java
@@ -46,8 +46,7 @@ public class UserService {
             throw new CustomException("동일한 아이디가 존재합니다");
         }
         // 1. db에 insert하기
-        int result = userRepository.insert(joinUserReqDto.getUsername(), joinUserReqDto.getPasswor(),
-                joinUserReqDto.getEmail());
+        int result = userRepository.insert();
 
         if (result != 1) {
             throw new CustomException("회원가입이 실패하였습니다");
