+++ src/main/resources/mapper/enterprise.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="shop.mtcoding.job.model.enterprise.EnterpriseRepository">
     <insert id="insert">
-      insert into enterprise_tb(enterprise_name, password, address, contact, image, email, size, sector, created_at) 
-      values(#{enterpriseName}, #{password}, #{address}, #{contact}, #{image}, #{email}, #{size}, #{sector}, now()) 
+      insert into enterprise_tb(enterprise_name, password, salt, address, contact, image, email, size, sector, created_at) 
+      values(#{enterpriseName}, #{password}, #{salt}, #{address}, #{contact}, #{image}, #{email}, #{size}, #{sector}, now()) 
    </insert>
     
    <update id="updateById">
